+++ cmd/dbaas/dbaas_list.go
@@ -15,10 +15,10 @@ import (
 )
 
 type dbaasServiceListItemOutput struct {
-	Name string `json:"name"`
-	Type string `json:"type"`
-	Plan string `json:"plan"`
-	Zone string `json:"zone"`
+	Name string      `json:"name"`
+	Type string      `json:"type"`
+	Plan string      `json:"plan"`
+	Zone v3.ZoneName `json:"zone"`
 }
 
 type dbaasServiceListOutput []dbaasServiceListItemOutput
@@ -32,7 +32,7 @@ type dbaasServiceListCmd struct {
 
 	_ bool `cli-cmd:"list"`
 
-	Zone string `cli-short:"z" cli-usage:"zone to filter results to"`
+	Zone v3.ZoneName `cli-short:"z" cli-usage:"zone to filter results to"`
 }
 
 func (c *dbaasServiceListCmd) CmdAliases() []string { return exocmd.GListAlias }
@@ -51,20 +51,12 @@ func (c *dbaasServiceListCmd) CmdPreRun(cmd *cobra.Command, args []string) error
 }
 
 func (c *dbaasServiceListCmd) CmdRun(_ *cobra.Command, _ []string) error {
-	var zones []v3.ZoneName
+	client := globalstate.EgoscaleV3Client
 	ctx := exocmd.GContext
 
-	if c.Zone != "" {
-		zones = []v3.ZoneName{v3.ZoneName(c.Zone)}
-	} else {
-		client, err := exocmd.SwitchClientZoneV3(ctx, globalstate.EgoscaleV3Client, v3.ZoneName(c.Zone))
-		if err != nil {
-			return err
-		}
-		zones, err = utils.AllZonesV3(ctx, client)
-		if err != nil {
-			return err
-		}
+	zones, err := utils.AllZonesV3(ctx, client, c.Zone)
+	if err != nil {
+		return err
 	}
 
 	out := make(dbaasServiceListOutput, 0)
@@ -77,14 +69,10 @@ func (c *dbaasServiceListCmd) CmdRun(_ *cobra.Command, _ []string) error {
 		}
 		done <- struct{}{}
 	}()
-	err := utils.ForEachZone(zones, func(zone v3.ZoneName) error {
-		ctx := exocmd.GContext
-		client, err := exocmd.SwitchClientZoneV3(ctx, globalstate.EgoscaleV3Client, v3.ZoneName(zone))
-		if err != nil {
-			return err
-		}
+	err = utils.ForEveryZone(zones, func(zone v3.Zone) error {
+		c := client.WithEndpoint((zone.APIEndpoint))
+		list, err := c.ListDBAASServices(ctx)
 
-		list, err := client.ListDBAASServices(ctx)
 		if err != nil {
 			return fmt.Errorf("unable to list Database Services in zone %s: %w", zone, err)
 		}
@@ -94,7 +82,7 @@ func (c *dbaasServiceListCmd) CmdRun(_ *cobra.Command, _ []string) error {
 				Name: string(dbService.Name),
 				Type: string(dbService.Type),
 				Plan: dbService.Plan,
-				Zone: string(zone),
+				Zone: zone.Name,
 			}
 		}